reimplement: SHC_3BB0A8C1_0x00512920 100.00% Reimplemented

+++ src/OpenSHC/Map/TileMapState/processMapOrientationChange.cpp
@@ -6,7 +6,7 @@ namespace Map {
     // FUNCTION: STRONGHOLDCRUSADER 0x00512920
     void TileMapState::processMapOrientationChange()
     {
-        if (this->DAT_FutureMapOrientation < 8) {
+        if ((int)this->DAT_FutureMapOrientation < 8) {
             MACRO_CALL_MEMBER(OpenSHC::Map::TileMapState_Func::changeMapOrientation, this)(
                 this->DAT_FutureMapOrientation);
             this->DAT_FutureMapOrientation = 8;
